fix(mcp): pass extraLinkDirs to linksCommand for MCP consistency
- MCP memex_links tool now reads config and passes extraLinkDirs,
  so agents get the same inbound link counts as CLI users
- Extract scanMarkdownFiles() into shared src/lib/scan.ts (DRY)
- Add self-reference guard: skip extraLinkDir if it resolves to cardsDir

Addresses review feedback on #123.
